+++ parkingLot/src/com/souvanik/parkinglot/strategy/BikePricing.java
@@ -1,5 +1,7 @@
 package com.souvanik.parkinglot.strategy;
 
+import java.math.BigDecimal;
+
 /*
  * Copyright (c) 2025 Souvanik Saha
  *
@@ -8,7 +10,7 @@ package com.souvanik.parkinglot.strategy;
  */
 public class BikePricing implements PricingStrategy{
     @Override
-    public double pricePerHour() {
-        return 10.0;
+    public BigDecimal pricePerHour() {
+        return BigDecimal.valueOf(10);
     }
 }
